csv-bigquery-loader/requirements.txt
google-cloud
 
google-cloud-bigquery
